Not found page added and folder renamed in camelcase
